refactor(v2-dto/phase0): 按业务域拆分 typings.d.ts + service.ts
- services/typings.d.ts → services/types/{common,cluster,frame,user,host,
  service,k8s,yarn,alarm,label,dag,deploy,addService}.d.ts
  TypeScript namespace merging 自动聚合，无需改 tsconfig
- services/service.ts 拆出 k8s.ts(7 函数) + yarn.ts(6 函数)
- 更新 6 个消费页面 import：K8s → k8s.ts，YARN → yarn.ts
- 验证：tsc --noEmit 零错误，lint 29 warning 均为预存
