+++ datasophon-ui-v2/src/pages/Cluster/ServiceInstance/Queue/BuildOrEditModal.tsx
@@ -7,7 +7,7 @@ import {
 } from '@ant-design/pro-components';
 import { message } from 'antd';
 import React from 'react';
-import { saveYarnQueue, updateYarnQueue } from '@/services/service';
+import { saveYarnQueue, updateYarnQueue } from '@/services/yarn';
 
 const SCHEDULE_POLICY_OPTIONS = [
   { label: 'fair', value: 'fair' },
