+++ datasophon-ui-v2/src/services/types/common.d.ts
@@ -1,0 +1,9 @@
+declare namespace DATASOPHON {
+  interface ApiResponse<T = any> {
+    success: boolean;
+    data: T;
+    errorCode?: number;
+    errorMessage?: string;
+    showType?: string;
+  }
+}
