+++ datasophon-ui-v2/src/services/types/frame.d.ts
@@ -1,0 +1,40 @@
+declare namespace DATASOPHON {
+  /** GET /v2/frame/list 响应体 */
+  interface FrameResponse {
+    id: number;
+    frameName: string;
+    frameCode: string;
+    frameVersion: string;
+  }
+
+  /** @deprecated 旧 DAO 镜像，迁移完成后删除，暂保留供未迁移文件使用 */
+  interface FrameInfo {
+    id: number;
+    frameName: string;
+    frameCode: string;
+    frameVersion: string;
+  }
+
+  interface FrameServiceItem {
+    id: number;
+    frameId: number;
+    frameCode: string;
+    serviceName: string;
+    serviceVersion: string;
+    serviceDesc?: string;
+  }
+
+  interface FrameK8sServiceItem {
+    id: number;
+    frameId: number;
+    serviceName: string;
+    serviceVersion: string;
+    serviceDesc?: string;
+    supportArtifacts?: string[];
+  }
+
+  interface FrameWithServices extends FrameInfo {
+    frameServiceList?: FrameServiceItem[];
+    frameK8sServiceList?: FrameK8sServiceItem[];
+  }
+}
