+++ datasophon-ui-v2/src/services/types/label.d.ts
@@ -1,0 +1,8 @@
+declare namespace DATASOPHON {
+  /** 节点标签，对应后端 ClusterNodeLabelEntity */
+  interface NodeLabel {
+    id: number;
+    nodeLabel: string;
+    clusterId?: number;
+  }
+}
