+++ datasophon-ui-v2/src/services/typings.d.ts
@@ -1,483 +1,2 @@
-declare namespace DATASOPHON {
-  interface ApiResponse<T = any> {
-    success: boolean;
-    data: T;
-    errorCode?: number;
-    errorMessage?: string;
-    showType?: string;
-  }
-  type ClusterArchType = 'physical' | 'k8s';
-
-  type ClusterState =
-    | '正在运行'
-    | '停止'
-    | '告警'
-    | '安装中'
-    | '停止中'
-    | '启动中'
-    | '未知';
-
-  interface ClusterManager {
-    id: number;
-    username: string;
-  }
-
-  // ── v2 DTO Request / Response（cluster 试点）────────────────────────────────
-
-  /** POST /v2/cluster 请求体 */
-  interface CreateClusterRequest {
-    clusterName: string;
-    clusterCode: string;
-    frameId: number;
-    archType: ClusterArchType;
-  }
-
-  /** PUT /v2/cluster/{id} 请求体（只修改名称和代号） */
-  interface UpdateClusterRequest {
-    clusterName: string;
-    clusterCode: string;
-  }
-
-  /** GET /v2/cluster/list、POST /v2/cluster 响应体 */
-  interface ClusterResponse {
-    id: number;
-    clusterName: string;
-    clusterCode: string;
-    clusterFrame?: string;
-    frameVersion?: string;
-    frameId?: number;
-    clusterState?: ClusterState;
-    clusterStateCode?: number;
-    archType?: ClusterArchType;
-    clusterManagerList?: ClusterManager[];
-  }
-
-  /** GET /v2/frame/list 响应体 */
-  interface FrameResponse {
-    id: number;
-    frameName: string;
-    frameCode: string;
-    frameVersion: string;
-  }
-
-  /** GET /v2/user/list 响应体（只含 id + username，不含 password） */
-  interface UserResponse {
-    id: number;
-    username: string;
-  }
-
-  interface ClusterInfo {
-    id: number;
-    clusterName: string;
-    clusterCode: string;
-    clusterFrame: string;
-    frameVersion?: string;
-    frameId?: number;
-    clusterState?: ClusterState;
-    clusterStateCode?: number;
-    archType?: ClusterArchType;
-    clusterManagerList?: ClusterManager[];
-  }
-
-  interface FrameInfo {
-    id: number;
-    frameName: string;
-    frameCode: string;
-    frameVersion: string;
-  }
-
-  interface HostInfo {
-    id: number;
-    hostname: string;
-    ip: string;
-    rack?: string;
-    coreNum?: number;
-    totalMem?: number;
-    totalDisk?: number;
-    usedMem?: number;
-    usedDisk?: number;
-    averageLoad?: string;
-    checkTime?: string;
-    clusterId?: number;
-    hostState: number;
-    cpuArchitecture?: string;
-    nodeLabel?: string;
-    serviceRoleNum?: number;
-    createTime?: string;
-    sshPort?: number;
-    sshUser?: string;
-  }
-
-  interface UserInfo {
-    id: number;
-    username: string;
-    email?: string;
-    phone?: string;
-    password?: string;
-    createTime?: string;
-    userType?: number;
-  }
-
-  interface CurrentUser extends UserInfo {
-    /** 'admin' | 'user' — used by access.ts for permission gating */
-    access?: string;
-    /** display name shown in ProLayout avatar */
-    name?: string;
-    /** avatar URL shown in ProLayout header */
-    avatar?: string;
-  }
-
-  interface ServiceInstanceInfo {
-    id: number;
-    clusterId: number;
-    serviceName: string;
-    label: string;
-    serviceState: string;
-    serviceStateCode: number;
-    needRestart: boolean;
-    frameServiceId: number;
-    sortNum: number;
-    dashboardUrl?: string;
-    alertNum: number;
-    catalog: string;
-    createTime?: string;
-    updateTime?: string;
-  }
-
-  interface ServiceRoleInstanceInfo {
-    id: number;
-    serviceId: number;
-    serviceRoleName: string;
-    hostname: string;
-    roleGroupId: number;
-    roleGroupName?: string;
-    serviceRoleState: string;
-    serviceRoleStateCode: number;
-    clusterId: number;
-    createTime?: string;
-  }
-
-  interface WebuiInfo {
-    name: string;
-    webUrl: string;
-  }
-
-  interface FrameServiceItem {
-    id: number;
-    frameId: number;
-    frameCode: string;
-    serviceName: string;
-    serviceVersion: string;
-    serviceDesc?: string;
-  }
-
-  interface FrameK8sServiceItem {
-    id: number;
-    frameId: number;
-    serviceName: string;
-    serviceVersion: string;
-    serviceDesc?: string;
-    supportArtifacts?: string[];
-  }
-
-  interface FrameWithServices extends FrameInfo {
-    frameServiceList?: FrameServiceItem[];
-    frameK8sServiceList?: FrameK8sServiceItem[];
-  }
-
-  // ── K8s 集群链路类型 ──────────────────────────────────────────────────────
-
-  /** K8s 集群 namespace（对应后端 K8sClusterNamespace） */
-  interface K8sNamespace {
-    id: number;
-    clusterId: number;
-    /** -1 未知 / 0 inactive / 1 active */
-    state: number;
-    namespace: string;
-  }
-
-  /** K8s 服务实例（对应后端 K8sServiceInstanceVO） */
-  interface K8sServiceInstanceVO {
-    id: number;
-    clusterId: number;
-    namespaceId: number;
-    namespace: string;
-    /** ENVIRONMENT / MIDDLEWARE / APPLICATION */
-    catalog: string;
-    serviceId: number;
-    serviceName: string;
-    /** 0 初始化 / 1 成功 / 2 失败 */
-    state: number;
-  }
-
-  /** K8s Helm values 完整记录（对应后端 K8sServiceInstanceValues） */
-  interface K8sInstanceValues {
-    id: number;
-    clusterId: number;
-    namespaceId: number;
-    serviceId: number;
-    instanceId: number;
-    /** 原始 yaml 文本（base values） */
-    values: string;
-    /** 用户新增的配置 yaml（delta values） */
-    deltaValues: string;
-    version: number;
-    /** 部署方式：helm / yaml */
-    metaFileType: string;
-  }
-
-  /** K8s Helm values 版本列表项（listSimpleByInstanceId 只返回部分字段） */
-  interface K8sInstanceValuesSimple {
-    id: number;
-    clusterId: number;
-    namespaceId: number;
-    serviceId: number;
-    instanceId: number;
-    version: number;
-  }
-
-  /** YARN 队列，对应后端 ClusterYarnQueue 实体 */
-  interface YarnQueue {
-    id?: number;
-    queueName: string;
-    minCore: number;
-    minMem: number;
-    maxCore: number;
-    maxMem: number;
-    appNum: number;
-    schedulePolicy: string;
-    weight: number;
-    /** 1 = 是，2 = 否 */
-    allowPreemption: number;
-    amShare: string;
-    clusterId?: number;
-    createTime?: string;
-  }
-
-  /**
-   * 服务配置参数项，对应后端 ServiceConfig 模型。
-   * type 决定前端渲染控件类型：
-   *   input / password / slider / switch / select / multipleSelect /
-   *   multiple / multipleWithKey / multipleWithMap
-   */
-  interface ConfigField {
-    name: string;
-    label: string;
-    value: any;
-    defaultValue?: any;
-    description?: string;
-    required: boolean;
-    enabled: boolean;
-    hidden?: boolean;
-    type: string;
-    configType?: string;
-    minValue?: number;
-    maxValue?: number;
-    unit?: string;
-    selectValue?: string[];
-    key?: string;
-    originalName?: string;
-    separator?: string;
-  }
-
-  /** 告警组，对应后端 AlertGroupEntity */
-  interface AlertGroup {
-    id?: number;
-    alertGroupName: string;
-    alertGroupCategory: string;
-    alertQuotaNum?: number;
-    clusterId?: number;
-    createTime?: string;
-  }
-
-  /** 告警指标，对应后端 ClusterAlertQuota 实体 */
-  interface AlertQuota {
-    id?: number;
-    alertQuotaName: string;
-    alertExpr: string;
-    compareMethod: string;
-    alertThreshold: number;
-    alertLevel: string;
-    alertGroupId: number;
-    alertGroupName?: string;
-    serviceRoleName: string;
-    noticeGroupId: number;
-    alertTactic: number;
-    intervalDuration: number;
-    triggerDuration: number;
-    alertAdvice: string;
-    quotaState?: string;
-    quotaStateCode?: number;
-  }
-
-  /** 节点标签，对应后端 ClusterNodeLabelEntity */
-  interface NodeLabel {
-    id: number;
-    nodeLabel: string;
-    clusterId?: number;
-  }
-
-  // ── DAG 命令（切片 6a/6b）────────────────────────────────────
-
-  /** DAG 状态五态，与后端 DagStatus 枚举对应 */
-  type DagStatus = 'PENDING' | 'RUNNING' | 'SUCCESS' | 'FAILED' | 'CANCEL';
-
-  /** 命令历史列表条目，对应后端 DagDefinitionEntity */
-  interface DagCommand {
-    id: string;
-    clusterId: number;
-    dagName: string;
-    description?: string;
-    status: DagStatus;
-    createdTime?: string;
-    startedTime?: string;
-    completedTime?: string;
-  }
-
-  /** DAG 节点，对应后端 InstallProgressDAG.nodes 中单个节点 */
-  interface DagNode {
-    id: string | number;
-    serviceInstanceId?: number;
-    clusterId?: number;
-    archType?: string;
-    status: DagStatus;
-    /** 节点显示名称 */
-    name?: string;
-    /** 主机名 */
-    hostname?: string;
-    /** 进度 0-100 */
-    progress?: number;
-    /** 命令 ID，用于查看日志 */
-    commandId?: string | number;
-    [key: string]: any;
-  }
-
-  /** DAG 边 */
-  interface DagEdge {
-    id: string | number;
-    start: string | number;
-    end: string | number;
-  }
-
-  /** DAG 图完整数据，对应后端 InstallProgressDAG */
-  interface DagGraph {
-    clusterId: number;
-    archType?: string;
-    nodes: DagNode[];
-    edges: DagEdge[];
-  }
-
-  // ─── 部署 (切片 7a) ───────────────────────────────────────────────────────
-
-  /** 上传临时文件返回值（UploadTempFile 实体） */
-  interface UploadedFile {
-    id: number;
-    fileName?: string;
-    filePath?: string;
-    fileSize?: number;
-    createTime?: string;
-  }
-
-  /** 校验部署清单返回值（ValidateResultVO） */
-  interface ValidateResult {
-    errors?: string[];
-    roles?: Array<{
-      serviceName: string;
-      version: string;
-      roleName: string;
-      deployHosts: string[];
-    }>;
-    k8sServices?: Array<{
-      serviceName: string;
-      version: string;
-      namespace: string;
-      metaFileType: string;
-    }>;
-  }
-
-  /** 执行部署返回值（InstallResult），含 dagId 用于跳转 DAG 图 */
-  interface DeployResult {
-    dagId: string;
-  }
-
-  // ─── 部署包 (切片 7b) ─────────────────────────────────────────────────────
-
-  /**
-   * 分片合并进度（MergeProgressVO）。
-   * state: 0=初始 | 2=处理中 | 1=完成 | -1=异常
-   * progress: 0-1 double，由后端计算 (merge+md5) / (total*2)
-   */
-  interface MergeProgress {
-    state: number;
-    error?: string;
-    total: number;
-    merge: number;
-    md5: number;
-    progress: number;
-  }
-
-  /**
-   * 导入组件进度（ImportCompProgressVO）。
-   * state: 0=初始化 | 2=解析元数据 | 3=解压安装包 | 4=保存数据 |
-   *        5=上传安装包到nexus | 6=上传镜像到nexus | 7=上传helm包到nexus |
-   *        1=成功 | -1=失败 | -2=进度对象不存在
-   */
-  interface ImportCompProgress {
-    progressId: number;
-    state: number;
-    error?: string;
-    total: number;
-    step: number;
-    progress: number;
-  }
-
-  /** 安装组件操作请求体 */
-  interface InstallComponentReq {
-    meteFileId: number;
-    pkgFileId?: number;
-    contentDecodePasswd: string;
-  }
-
-  // ─── 添加服务向导 (切片 7c) ───────────────────────────────────────────────
-
-  /** 部署清单上下文（步骤 1 产出，后续步骤复用） */
-  interface ManifestContext {
-    deployFileId: number;
-    contentDecodePasswd: string;
-  }
-
-  /** 框架服务（FrameServiceEntity），list-newest 返回 */
-  interface FrameService {
-    id: number;
-    serviceName: string;
-    label?: string;
-    serviceVersion?: string;
-    serviceDesc?: string;
-    frameCode?: string;
-    /** 是否已安装到集群 */
-    installed?: boolean;
-    /** 部署清单中是否勾选 */
-    selected?: boolean;
-  }
-
-  /**
-   * 框架服务角色（FrameServiceRoleEntity）。
-   * serviceRoleType: master | worker | client
-   * cardinality: "1" 单选 | "1+"/"N+" 多选
-   */
-  interface FrameServiceRole {
-    id: number;
-    serviceId: number;
-    serviceRoleName: string;
-    serviceRoleType: string;
-    cardinality?: string;
-    /** 部署清单回填的主机列表 */
-    hosts?: string[];
-  }
-
-  /** 服务角色与主机映射（提交给 role-host-mapping） */
-  interface RoleHostMapping {
-    serviceRole: string;
-    hosts: string[];
-  }
-}
+// 类型已按业务域拆分至 services/types/<domain>.d.ts
+// TypeScript 通过 declare namespace DATASOPHON {} 声明合并自动聚合，无需手动引入。
